Rename the handlers and views

+++ article.go
@@ -7,6 +7,7 @@ import (
 
 	sessions "github.com/goincremental/negroni-sessions"
 	"github.com/julienschmidt/httprouter"
+	"github.com/russross/blackfriday"
 )
 
 // Article is the type that holds blog posts.
@@ -23,7 +24,14 @@ type Article struct {
 	PublishedAt time.Time `db:"published_at"`
 }
 
-func ArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+// MarshalHTML returns the text of the article in HTML format.
+func (a Article) MarshalHTML() string {
+	return blackfriday.MarkdownCommon([]byte(a.Text))
+}
+
+// View handle the displaying of articles.
+func View(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+
 	var article Article
 
 	err := database.Get(&article, "SELECT * FROM articles WHERE slug = ?", p.ByName("slug"))
@@ -36,6 +44,7 @@ func ArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params)
 		return
 	}
 
+	// Reject the request if the user isn't logged and that the article isn't published.
 	if sessions.GetSession(r).Get("logged") != true && !article.IsPublished {
 		http.Error(w, "Unauthorized", http.StatusUnauthorized)
 		return
@@ -47,7 +56,16 @@ func ArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params)
 	})
 }
 
-func DeleteArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+// Delete remove articles.
+func Delete(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+
+	// Ensure the user is logged.
+	if sessions.GetSession(r).Get("logged") != true {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
+	// Remote the article from the database.
 	_, err := database.Exec("DELETE FROM articles WHERE slug = ?", p.ByName("slug"))
 	if err != nil {
 		panic(err)
@@ -56,9 +74,18 @@ func DeleteArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.P
 	http.Redirect(w, r, "/read", http.StatusFound)
 }
 
-func EditArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+// Edit show the edition form for articles.
+func Edit(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+
+	// Ensure the user is logged
+	if sessions.GetSession(r).Get("logged") != true {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	var article Article
 
+	// Get the article from the database.
 	err := database.Get(&article, "SELECT * FROM articles WHERE slug = ?", p.ByName("slug"))
 	if err != nil && err != sql.ErrNoRows {
 		panic(err)
@@ -75,7 +102,15 @@ func EditArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Par
 	})
 }
 
-func EditArticleFormHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+// Update do the actual modification of the article in the database.
+func Update(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+
+	// Ensure the user is logged.
+	if sessions.GetSession(r).Get("logged") != true {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	var article Article
 
 	err := database.Get(&article, "SELECT * FROM articles WHERE slug = ?", p.ByName("slug"))
@@ -88,6 +123,7 @@ func EditArticleFormHandler(w http.ResponseWriter, r *http.Request, p httprouter
 		return
 	}
 
+	// Validate the request's form data for the article.
 	var v = NewValidator(r)
 	v.NotEmpty("title")
 	v.MaxLen("title", 150)
@@ -98,6 +134,7 @@ func EditArticleFormHandler(w http.ResponseWriter, r *http.Request, p httprouter
 	v.MaxLen("tagline", 450)
 	v.NotEmpty("text")
 
+	// In case of error, forward them to the next request.
 	if v.HasErrors() {
 		sessions.GetSession(r).AddFlash(v.Errors(), "_errors")
 		sessions.GetSession(r).AddFlash(r.Form, "_inputs")
@@ -105,6 +142,7 @@ func EditArticleFormHandler(w http.ResponseWriter, r *http.Request, p httprouter
 		return
 	}
 
+	// If everything is fine, update the article in the database.
 	_, err = database.Exec("UPDATE articles SET title = ?, slug = ?, tagline = ?, text = ?, tags = ?, updated_at = datetime('now') WHERE id = ?", r.FormValue("title"), r.FormValue("slug"), r.FormValue("tagline"), r.FormValue("text"), r.FormValue("tags"), article.ID)
 	if err != nil {
 		panic(err)
@@ -114,7 +152,16 @@ func EditArticleFormHandler(w http.ResponseWriter, r *http.Request, p httprouter
 	return
 }
 
-func PublishArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+// Publish set the published flag of an article to true.
+func Publish(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+
+	// Ensure the user is logged.
+	if sessions.GetSession(r).Get("logged") != true {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
+	// Switch the flag on.
 	_, err := database.Exec("UPDATE articles SET is_published = ?, published_at = datetime('now') WHERE slug = ?", true, p.ByName("slug"))
 	if err != nil {
 		panic(err)
@@ -123,10 +170,22 @@ func PublishArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.
 	http.Redirect(w, r, r.Referer(), http.StatusFound)
 }
 
-func ReadHandler(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+// List show all articles available to the user.
+func List(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+
 	var articles []*Article
 
-	err := database.Select(&articles, "SELECT * FROM articles ORDER BY published_at DESC")
+	// Get the login status of the user.
+	logged := sessions.GetSession(r).Get("logged").(bool)
+
+	// If the user isn't logged, only show published articles and order them by publication date.
+	// If the user is logged, show all articles and order them by creation date.
+	var err error
+	if logged {
+		err = database.Select(&articles, "SELECT * FROM articles WHERE is_published = true ORDER BY published_at DESC")
+	} else {
+		err = database.Select(&articles, "SELECT * FROM articles WHERE ORDER BY created_at DESC")
+	}
 	if err != nil {
 		panic(err)
 	}
@@ -134,10 +193,20 @@ func ReadHandler(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
 	render(w, r, "read", map[string]interface{}{
 		"Title":    "Read",
 		"Articles": articles,
+		"Logged":   logged,
 	})
 }
 
-func UnpublishArticleHandler(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+// Unpublish switch off visibility of an article.
+func Unpublish(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+
+	// Ensure the user is logged.
+	if sessions.GetSession(r).Get("logged") != true {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
+	// Switch the flag off.
 	_, err := database.Exec("UPDATE articles SET is_published = ? WHERE slug = ?", false, p.ByName("slug"))
 	if err != nil {
 		panic(err)
@@ -146,13 +215,30 @@ func UnpublishArticleHandler(w http.ResponseWriter, r *http.Request, p httproute
 	http.Redirect(w, r, r.Referer(), http.StatusFound)
 }
 
-func WriteHandler(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+// Write display the form for writing articles.
+func Write(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+
+	// Ensure the user is logged.
+	if sessions.GetSession(r).Get("logged") != true {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	render(w, r, "write", map[string]interface{}{
 		"Title": "Write",
 	})
 }
 
-func WriteFormHandler(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+// Create validate and add new articles in the database.
+func Create(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+
+	// Ensure the user is logged.
+	if sessions.GetSession(r).Get("logged") != true {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
+	// Validate the request form data.
 	var v = NewValidator(r)
 	v.NotEmpty("title")
 	v.MaxLen("title", 150)
@@ -163,6 +249,8 @@ func WriteFormHandler(w http.ResponseWriter, r *http.Request, _ httprouter.Param
 	v.MaxLen("tagline", 450)
 	v.NotEmpty("text")
 
+	// In case of error, return the user to the previous page with a listing of
+	// the errors and inputs.
 	if v.HasErrors() {
 		sessions.GetSession(r).AddFlash(v.Errors(), "_errors")
 		sessions.GetSession(r).AddFlash(r.Form, "_inputs")
@@ -170,11 +258,11 @@ func WriteFormHandler(w http.ResponseWriter, r *http.Request, _ httprouter.Param
 		return
 	}
 
+	// Insert the new article in the database.
 	_, err := database.Exec("INSERT INTO articles (title, slug, tagline, text, tags, is_published, created_at, updated_at, published_at) VALUES (?, ?, ?, ?, ?, ?, datetime('now'), datetime('now'), datetime('now'))", r.FormValue("title"), r.FormValue("slug"), r.FormValue("tagline"), r.FormValue("text"), r.FormValue("tags"), false)
 	if err != nil {
 		panic(err)
 	}
 
 	http.Redirect(w, r, "/article/"+r.FormValue("slug"), http.StatusFound)
-	return
 }
